backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.0
google-generativeai==0.3.0
twscrape==0.4.0
aiohttp==3.9.1
python-dotenv==1.0.0
schedule==1.2.0
playwright==1.40.0

